% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fosr_vs.R
\name{fosr_vs}
\alias{fosr_vs}
\title{Ported function from \code{'refund'} package}
\usage{
fosr_vs(
  formula,
  data,
  nbasis = 10,
  method = c("ls", "grLasso", "grMCP", "grSCAD"),
  epsilon = 1e-05,
  max.iter_num = 100,
  groups = NULL
)
}
\arguments{
\item{formula, data, nbasis, method, epsilon, max.iter_num}{see
\code{\link[refund]{fosr.vs}}}

\item{groups}{integer vector with length of number of time-points
of how time-points should be grouped; default
is \code{NULL}, indicating there is no local sparsity.}
}
\description{
A modified version of \code{\link[refund]{fosr.vs}}, but
with groups parameter to allow grouping time points rather than the
whole coefficient when the underlying functions are locally supported.
}
