% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spfda_simulate}
\alias{spfda_simulate}
\title{Generate toy example data}
\usage{
spfda_simulate(n = 1000, n_timepoints = 100, err = 1, seed = 1)
}
\arguments{
\item{n}{Total number of observations}

\item{n_timepoints}{Total number of time points}

\item{err}{error level}

\item{seed}{random seed}
}
\value{
A list of data generated
}
\description{
Generate toy example data
}
