% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{spfda_weight}
\alias{spfda_weight}
\title{Calculates weight matrices}
\usage{
spfda_weight(X, Y, bandwidth, part)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response matrix}

\item{bandwidth}{numeric band-width}

\item{part}{list of time point boundaries}
}
\value{
the weight matrix
}
\description{
Calculates weight matrices
}
